pyproject.toml
[project]
name = "ad-inventory-forecast"
version = "0.1.0"
description = "Ad inventory demand forecasting system using BigQuery ML"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.25.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["config", "src"]
